+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * don't force `as.numeric` when unnecessary; #182 addressing #181
 
+* new method `drop_units` for data frames; #191 addressing #187
+
 # version 0.6-2
 
 * fix support for logarithms and decibels; #177 addressing #176