+++ NEWS.md
@@ -1,3 +1,7 @@
+# version 0.6-4
+
+* fix support for weights with units in `weighted.mean`; #205
+
 # version 0.6-3
 
 * improve platform dependent encodings handling; #183